fix add product
Fixes add product form by deleten unnecessary functions to set the
values of the form fields, as these functions did't have any type of
field validation or value modification.

With this fix the Product's attributes are read from the forms fields on
form submission.

Fixes the product id issue by assigning the lenght of the product array
as the next id.

Removes unused states and imports.